Refactor authz_geoserver_test.py
