+++ src/layman/gs_wfs_proxy.py
@@ -1,11 +1,8 @@
-from urllib.parse import urljoin
 import requests
 
-from flask import Blueprint, g, current_app as app, request, Response, jsonify
+from flask import Blueprint, g, current_app as app, request, Response
 
-from layman.authn import authenticate, flush_cache
-from layman.authz import authorize
-
+from layman.authn import authenticate
 from layman import settings
 
 bp = Blueprint('gs_wfs_proxy_bp', __name__)
@@ -21,21 +18,12 @@ def before_request():
 def proxy(subpath):
     app.logger.info(f"GET WFS proxy, user={g.user}, subppath={subpath}, url={request.url}, request.query_string={request.query_string.decode('UTF-8')}")
 
-# TODO
-# [x]    1. headers
-# [x]    2. data
-# [x]    3. cookies
-# [x]    4. url
-# [x]    5. username
-# [ ]    6. auth
-
     url = settings.LAYMAN_GS_URL + subpath + '?' + request.query_string.decode('UTF-8')
     headers_req = {key.lower(): value for (key, value) in request.headers if key != 'Host'}
     if g.user is not None:
         headers_req[settings.LAYMAN_GS_AUTHN_HTTP_HEADER_ATTRIBUTE] = g.user['username']
-    # headers_req[settings.LAYMAN_GS_AUTHN_HTTP_HEADER_ATTRIBUTE] = "layman"
 
-    app.logger.info(f"GET WFS proxy, headers_req={headers_req}")
+    app.logger.info(f"GET WFS proxy, headers_req={headers_req}, url={url}")
     response = requests.request(method=request.method,
                                 url=url,
                                 data=request.get_data(),
